integration of graph-mvvm-db
