+++ Abreath/.idea/misc.xml
@@ -4,16 +4,17 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="../../../../../layout/custom_preview.xml" value="0.3255208333333333" />
-        <entry key="app/src/main/res/drawable/gradient_header.xml" value="0.2115" />
+        <entry key="app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.303" />
+        <entry key="app/src/main/res/drawable/gradient_header.xml" value="0.216" />
         <entry key="app/src/main/res/drawable/ic_baseline_settings_24.xml" value="0.384" />
         <entry key="app/src/main/res/drawable/ic_dashboard_black_24dp.xml" value="0.303" />
         <entry key="app/src/main/res/drawable/ic_home_black_24dp.xml" value="0.29" />
         <entry key="app/src/main/res/drawable/ic_launcher_background.xml" value="0.303" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.1" />
         <entry key="app/src/main/res/layout/fragment_dashboard.xml" value="0.33" />
-        <entry key="app/src/main/res/layout/fragment_home.xml" value="0.25" />
+        <entry key="app/src/main/res/layout/fragment_home.xml" value="0.1" />
         <entry key="app/src/main/res/layout/fragment_notifications.xml" value="0.5359375" />
-        <entry key="app/src/main/res/layout/fragment_profile_graph.xml" value="0.23125" />
+        <entry key="app/src/main/res/layout/fragment_profile_graph.xml" value="0.2994791666666667" />
         <entry key="app/src/main/res/layout/fragment_settings.xml" value="0.2994791666666667" />
         <entry key="app/src/main/res/menu/bottom_nav_menu.xml" value="0.2994791666666667" />
       </map>
