+++ Abreath/app/src/main/java/com/coen390/abreath/MainActivity.java
@@ -22,6 +22,7 @@ import com.coen390.abreath.databinding.ActivityMainBinding;
 
 import java.util.Objects;
 
+
 public class MainActivity extends AppCompatActivity {
 
     private ActivityMainBinding binding;
@@ -41,6 +42,9 @@ public class MainActivity extends AppCompatActivity {
         AppBarConfiguration appBarConfiguration = new AppBarConfiguration.Builder(
                 R.id.navigation_home, R.id.navigation_dashboard, R.id.navigation_settings)
                 .build();
+
+        getSupportActionBar().setElevation(0f);
+
 
 
         NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment_activity_main);
