--- Abreath/app/src/main/java/com/coen390/abreath/ui/home/ProfileGraph.java
@@ -1,112 +1,0 @@
-package com.coen390.abreath.ui.home;
-
-import android.graphics.Color;
-import android.os.Bundle;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import com.coen390.abreath.R;
-import com.github.mikephil.charting.charts.BarChart;
-import com.github.mikephil.charting.components.XAxis;
-import com.github.mikephil.charting.components.YAxis;
-import com.github.mikephil.charting.data.BarData;
-import com.github.mikephil.charting.data.BarDataSet;
-import com.github.mikephil.charting.data.BarEntry;
-import com.github.mikephil.charting.formatter.ValueFormatter;
-import com.github.mikephil.charting.interfaces.datasets.IBarDataSet;
-
-import java.util.ArrayList;
-import java.util.Random;
-
-/**
- * A simple {@link Fragment} subclass.
- * Use the {@link ProfileGraph#newInstance} factory method to
- * create an instance of this fragment.
- */
-
-//https://learntodroid.com/how-to-display-a-bar-chart-in-your-android-app/#:~:text=To%20display%20a%20bar%20chart%20in%20your%20Android%20app%20you,appearance%20of%20the%20bar%20chart
-
-
-public class ProfileGraph extends Fragment {
-
-
-    private static final int MAX_X_VALUE = 7;
-    private static final int MAX_Y_VALUE = 50;
-    private static final int MIN_Y_VALUE = 5;
-    private static final String SET_LABEL = "App Downloads";
-    private static final String[] DAYS = { "SUN", "MON", "TUE", "WED", "THU", "FRI", "SAT" };
-
-    private BarChart chart;
-
-    private BarData createChartData() {
-        ArrayList<BarEntry> values = new ArrayList<>();
-        for (int i = 0; i < MAX_X_VALUE; i++) {
-            float x = i;
-            // float y = new Util().randomFloatBetween(MIN_Y_VALUE, MAX_Y_VALUE);
-            float y = MIN_Y_VALUE + new Random().nextFloat() * (MAX_Y_VALUE - MIN_Y_VALUE);
-            values.add(new BarEntry(x, y));
-        }
-
-        BarDataSet set1 = new BarDataSet(values, SET_LABEL);
-
-        ArrayList<IBarDataSet> dataSets = new ArrayList<>();
-        dataSets.add(set1);
-
-        BarData data = new BarData(dataSets);
-
-        return data;
-    }
-
-    @Override
-    public void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-    }
-
-    @Nullable
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_profile_graph, container, false);
-
-        chart = view.findViewById(R.id.fragment_verticalbarchart_chart);
-
-        BarData data = createChartData();
-        configureChartAppearance();
-        prepareChartData(data);
-
-        return view;
-    }
-    private void prepareChartData(BarData data) {
-        data.setValueTextSize(12f);
-        chart.setData(data);
-        chart.invalidate();
-    }
-
-    private void configureChartAppearance() {
-        chart.getDescription().setEnabled(false);
-        chart.setDrawValueAboveBar(false);
-
-        XAxis xAxis = chart.getXAxis();
-        xAxis.setValueFormatter(new ValueFormatter() {
-            @Override
-            public String getFormattedValue(float value) {
-                return DAYS[(int) value];
-            }
-        });
-
-        YAxis axisLeft = chart.getAxisLeft();
-        axisLeft.setGranularity(10f);
-        axisLeft.setAxisMinimum(0);
-        axisLeft.setAxisLineColor(Color.BLUE);
-
-        YAxis axisRight = chart.getAxisRight();
-        axisRight.setGranularity(10f);
-        axisRight.setAxisMinimum(0);
-    }
-
-}
